% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deerdat.R
\docType{data}
\name{deerdat}
\alias{deerdat}
\title{Hunter-killed deer in New Jersey}
\format{
A dataframe with 36 rows and 3 variables
}
\description{
Data on hunter-killed deer in New Jersey.
}
\details{
\itemize{
		\item county. New Jersey county.  Hunterdon, Warren, and Hudson counties are not included.
		\item yr. Year of data.  Either 1997 or 2007.
		\item hk. Number of hunter-killed deer per square mile.
}
}
\keyword{datasets}
