% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deerdatw.R
\docType{data}
\name{deerdatw}
\alias{deerdatw}
\title{Hunter-killed deer in New Jersey}
\format{
A dataframe with 18 rows and 3 variables
}
\description{
Data on hunter-killed deer in New Jersey.
}
\details{
\itemize{
		\item county. New Jersey county.  Hunterdon, Warren, and Hudson counties are not included.
		\item y1997. Number of hunter-killed deer per square mile in 1997.
		\item y2007 Number of hunter-killed deer per square mile in 2007
}
}
\keyword{datasets}
