% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevdat.R
\docType{data}
\name{fevdat}
\alias{fevdat}
\title{FEV data}
\format{
A dataframe with 654 rows and 6 variables
}
\description{
A dataset of forced expiratory volume from Rosner, B. (1999),
Fundamentals of Biostatistics, 5th ed., Pacific Grove, CA: Duxbury.
Data obtained from http://biostat.mc.vanderbilt.edu/DataSets
 The variables are:
}
\details{
\itemize{
		\item id. Participant ID
		\item age. Age in years
		\item fev. Forced expiratory volume in liters
		\item height. Height in inches
		\item sex. Sex assigned at birth (male, female)
		\item smoke. Current smoking status (current smoker, non-current smoker)
	...
}
}
\keyword{datasets}
