% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevdat10.R
\docType{data}
\name{fevdat10}
\alias{fevdat10}
\title{FEV data for 10 year olds}
\format{
A dataframe with 654 rows and 6 variables
}
\description{
A dataset of forced expiratory volume for 10 year olds from Rosner, B. (1999),
Fundamentals of Biostatistics, 5th ed., Pacific Grove, CA: Duxbury.
Data obtained from http://biostat.mc.vanderbilt.edu/DataSets
 The variables are:
}
\details{
\itemize{
		\item id. Participant ID
		\item fev. Forced expiratory volume in liters
		\item height. Height in inches
		\item sex. Sex assigned at birth (male, female)
		\item smoke. Current smoking status (current smoker, non-current smoker)
	...
}
}
\keyword{datasets}
