% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\docType{data}
\name{support}
\alias{support}
\title{SUPPORT Phase I}
\format{
A dataframe with 651 rows and 5 variables
}
\description{
Limited data from Phase I of the SUPPORT study (Study to Understand Prognoses
Preferences Outcomes and Risks of Treatment).
}
\details{
Knaus WA, Harrell FE, Lynn J et al. (1995): The SUPPORT prognostic model: Objective estimates of survival for seriously ill hospitalized adults. Annals of Internal Medicine 122:191-203.
Data obtained from http://biostat.mc.vanderbilt.edu/DataSets

\itemize{
		\item age. Age in years
		\item death. Binary indicator variable of death (1 = dead).
		\item dzclass. Diagnosis class of: ARF/MOSF (acute renal failure or multiple organ system failure) COPD/CHF/Cirrhosis (chronic obstructive pulmonary disease, congestive heart failure, or cirrhosis), coma, or cancer.
		\item income. Categorical income variable.
		\item temp. Temperature in degrees Celcius.
	...
}
}
\keyword{datasets}
