% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevdat10.R, R/support.R
\docType{data}
\name{fevdat10}
\alias{fevdat10}
\title{FEV data for 10 year olds}
\format{
A dataframe with 654 rows and 6 variables

A dataframe with 654 rows and 6 variables
}
\description{
A dataset of forced expiratory volume for 10 year olds from Rosner, B. (1999),
Fundamentals of Biostatistics, 5th ed., Pacific Grove, CA: Duxbury.
Data obtained from http://biostat.mc.vanderbilt.edu/DataSets
 The variables are:

Limited data from Phase I of the SUPPORT study (Study to Understand Prognoses
Preferences Outcomes and Risks of Treatment).
Knaus WA, Harrell FE, Lynn J et al. (1995): The SUPPORT prognostic model: Objective estimates of survival for seriously ill hospitalized adults. Annals of Internal Medicine 122:191-203.
Data obtained from http://biostat.mc.vanderbilt.edu/DataSets
 The variables are:
}
\details{
\itemize{
		\item id. Participant ID
		\item fev. Forced expiratory volume in liters
		\item height. Height in inches
		\item sex. Sex assigned at birth (male, female)
		\item smoke. Current smoking status (current smoker, non-current smoker)
	...
}

\itemize{
		\item age. Age in years
		\item death. Binary indicator variable of death (1 = dead).
		\item dzclass. Diagnosis class of: ARF/MOSF (acute renal failure or multiple organ system failure) COPD/CHF/Cirrhosis (chronic obstructive pulmonary disease, congestive heart failure, or cirrhosis), coma, or cancer.
		\item income. Categorical income variable.
		\item temp. Temperature in degrees Celcius.
	...
}
}
\keyword{datasets}
